+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * use object for option
-* `.fit()` to `.applySchema()` / `.applyTo()`
+* `.fit()` to `.applySchemaTo()` / `.applyTo()`
 
 ## [2.2.6] - 2020-02-14